PortalNews/requirements.txt
Django>=5.0
django-filter
